Refactor launcher CLI into modular commands
- Split command handling into per-command modules
- Inject launcher invocation formatting for help and errors
